refactor: 코드 정리 #4
- 사용하지 않는 코드 삭제
- 잘못된 패키지 위치 수정
